requirements.txt
apipkg==1.5
astroid==2.4.0
attrs==19.3.0
click==7.1.1
dnspython==1.16.0
execnet==1.7.1
Flask==1.1.2
Flask-MySQLdb==0.2.0
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.2
packaging==20.3
pep8==1.7.1
pluggy==0.13.1
protobuf==3.6.1
py==1.8.1
pylint==2.5.0
pyparsing==2.4.7
pytest==5.4.1
pytest-cache==1.0
pytest-pep8==1.0.6
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
vsearch==1.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
